// packages/unsaged/utils/app/storage/supabase/types/supabase.ts
export interface SupaDatabase {
  public: {
    Tables: {
// ...
      system_prompts: {
        Row: {
          content: string;
          folder_id: string | null;
          id: string;
          models: string[];
          name: string;
          user_id: string;
        };
        Insert: {
          content: string;
          folder_id?: string | null;
          id?: string;
          models: string[];
          name: string;
          user_id?: string;
        };
        Update: {
          content?: string;
          folder_id?: string | null;
          id?: string;
          models?: string[];
          name?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'system_prompt_owner';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'system_prompt_owner_folder';
            columns: ['folder_id'];
            isOneToOne: false;
            referencedRelation: 'folders';
            referencedColumns: ['id'];
          },
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
}